first_semester/lab7/app/app.component.ts
import {Component, OnInit} from '@angular/core';
import { CommonModule } from '@angular/common';
import { RouterOutlet } from '@angular/router';
import {User} from "./app.User";
import {FormBuilder, FormGroup, ReactiveFormsModule, Validators} from "@angular/forms";

@Component({
  selector: 'app-root',
  standalone: true,
  imports: [CommonModule, RouterOutlet, ReactiveFormsModule],
  templateUrl: './app.component.html',
  styleUrl: './app.component.css'
})
export class AppComponent implements OnInit{
  userForm!: FormGroup;
  user: User = new User();
  title: string = '7';

  constructor(private fb: FormBuilder) {}

  ngOnInit(): void {
    this.initForm();
  }

  private initForm() {
    this.userForm = this.fb.group({
      login: ['', Validators.required],
      password: ['', Validators.required],
      confirmPassword: ['', Validators.required],
      fullName: ['', Validators.required],
      email: ['', [Validators.required, Validators.email]],
      homepage: [''],
      icqNumber: ['', Validators.pattern(/^\d+$/)],
      skypeNickName: [''],
    });
  }

  saveForm() {
    if (this.userForm.valid) {
      this.user = { ...this.userForm.value };
      console.log('Form saved:', this.user);
      this.userForm.reset();
    } else {
      this.markFormGroupTouched(this.userForm);
    }
  }
 markFormGroupTouched(formGroup: FormGroup) {
    Object.values(formGroup.controls).forEach(control => {
      if (control instanceof FormGroup) {
        this.markFormGroupTouched(control);
      } else {
        control.markAsTouched();
      }
    });
  }
}
